move cross-section visual to viz.py
